Logros

+++ app/Http/Controllers/ReservaController.php
@@ -5,9 +5,36 @@ namespace App\Http\Controllers;
 use App\Models\Destino;
 use App\Models\Reserva;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ReservaController extends Controller
 {
+    /**
+     * Dirige a la vista donde se muestran todas las reservas existentes. Panel Admin.
+     *
+     * @return void
+     */
+    public function index()
+    {
+        $reservas = Reserva::all()->sortBy('created_at');
+        return view('admin.reservas.index', [
+            'reservas' => $reservas,
+        ]);
+    }
+
+    /**
+     * Muestra en una vista los detalles de una reserva.
+     *
+     * @param  mixed $reserva
+     * @return void
+     */
+    public function detallesReserva(Reserva $reserva)
+    {
+        return view('admin.reservas.detalles', [
+            'reserva' => $reserva,
+        ]);
+    }
+
     /**
      * Crea una reserva nueva con el usuario actualmente logeado, obtiene los datos de la reserva por
      * método POST.
@@ -28,6 +55,10 @@ class ReservaController extends Controller
 
         $mail = new MailController();
         $mail->index($n_reserva);
+
+        if ($request->precioAct == 0) {
+            return redirect()->route('usuarios.index')->with('success', '¡La reserva se ha creado correctamente!');
+        }
     }
 
     /**
@@ -44,7 +75,7 @@ class ReservaController extends Controller
             ->get();
         $destinos = Destino::select('nombre', 'comarca')->get();
         $reservas = Reserva::where('user_id', $request->user()->id)
-            ->orderBy('fecha', 'DESC')
+            ->orderBy('created_at', 'DESC')
             ->paginate(5);
 
         return view('gadiritas.reservas', compact('reservas', 'comarcas', 'destinos'));
@@ -58,10 +89,15 @@ class ReservaController extends Controller
      */
     public function borrarReserva(Request $request)
     {
-        $reserva = $request->input('id');
-        Reserva::where('id', $reserva)->delete();
-        $mail = new MailController();
-        $mail->cancelar();
-        return redirect('/reservas');
+        if (Auth::user()) {
+            $reserva = $request->input('id');
+            Reserva::where('id', $reserva)->delete();
+            $mail = new MailController();
+            $mail->cancelar();
+            if (Auth::user()->is_admin) {
+                return redirect('/reservas');
+            }
+            return redirect('/user/reservas');
+        }
     }
 }
